Adding proper 3d and 4d tests. Split test files
Conflicts:
	tests/unit/tests_high_five_base.cpp
