+++ CMake/portability/CheckCXXStandardSupport.cmake
@@ -22,3 +22,14 @@ endif()
 check_cxx_compiler_flag("${CMAKE_CXX11_STANDARD_COMPILE_OPTION}" COMPILER_SUPPORTS_CXX11)
 check_cxx_compiler_flag("${CMAKE_CXX14_STANDARD_COMPILE_OPTION}" COMPILER_SUPPORTS_CXX14)
 check_cxx_compiler_flag("${CMAKE_CXX17_STANDARD_COMPILE_OPTION}" COMPILER_SUPPORTS_CXX17)
+
+function(select_recent_cxx_standard_available)
+    if(CMAKE_CXX_STANDARD)
+        return()
+    endif()
+    if(COMPILER_SUPPORTS_CXX14)
+        set(CMAKE_CXX_STANDARD 14 PARENT_SCOPE)
+    elseif(COMPILER_SUPPORTS_CXX11)
+        set(CMAKE_CXX_STANDARD 11 PARENT_SCOPE)
+    endif()
+endfunction()
