+++ tests/unit/CMakeLists.txt
@@ -8,8 +8,8 @@ if(NOT Boost_USE_STATIC_LIBS)
   add_definitions(-DBOOST_TEST_DYN_LINK=TRUE)
 endif()
 
-## base tests
-foreach(test_name tests_high_five_base tests_high_five_easy)
+
+foreach(test_name tests_high_five_base tests_high_five_multi_dims tests_high_five_easy)
   add_executable(${test_name} "${test_name}.cpp")
   target_link_libraries(${test_name}
                         ${Boost_UNIT_TEST_FRAMEWORK_LIBRARIES}
@@ -35,7 +35,7 @@ if(HIGHFIVE_PARALLEL_HDF5)
   target_link_libraries(tests_parallel_bin
     ${Boost_UNIT_TEST_FRAMEWORK_LIBRARIES} ${HDF5_C_LIBRARIES}
     ${MPI_C_LIBRARIES})
-  add_test(NAME tests_parallel COMMAND ${TEST_MPI_EXEC_PREFIX_DEFAULT} -n 8
+  add_test(NAME tests_parallel COMMAND ${TEST_MPI_EXEC_PREFIX_DEFAULT}
     ${CMAKE_CURRENT_BINARY_DIR}/tests_parallel_bin)
 
 endif()
